+++ src/test/java/com/kputimcevs/insurance/premium/PremiumCalculatorTest.java
@@ -25,35 +25,35 @@ class PremiumCalculatorTest {
     @Test
     public void premiumCalculatorResolvesCorrectValue() {
         float result = premiumCalculator.calculate(policy);
-        float expected = (float) 2.28;
+        float expected = 2.28f;
         assertEquals(expected, result);
     }
 
     @Test
     public void premiumCalculatorResolvesCorrectValueWhenFireRiskExceedsThreshold() {
-        policy.policyObjects.get(0).subObjects.get(0).sumInsured = (float) 20.5;
+        policy.policyObjects.get(0).subObjects.get(0).sumInsured = 20.5f;
 
         float result = premiumCalculator.calculate(policy);
-        float expected = (float) 3.28;
+        float expected = 3.28f;
         assertEquals(expected, result);
     }
 
     @Test
     public void premiumCalculatorResolvesCorrectValueWhenTheftRiskExceedsThreshold() {
-        policy.policyObjects.get(0).subObjects.get(1).sumInsured = (float) 9.48;
+        policy.policyObjects.get(0).subObjects.get(1).sumInsured = 9.48f;
 
         float result = premiumCalculator.calculate(policy);
-        float expected = (float) 2.15;
+        float expected = 2.15f;
         assertEquals(expected, result);
     }
 
     @Test
     public void premiumCalculatorResolvesCorrectValueWhenBothRisksExceedsThreshold() {
-        policy.policyObjects.get(0).subObjects.get(0).sumInsured = (float) 20.5;
-        policy.policyObjects.get(0).subObjects.get(1).sumInsured = (float) 9.48;
+        policy.policyObjects.get(0).subObjects.get(0).sumInsured = 20.5f;
+        policy.policyObjects.get(0).subObjects.get(1).sumInsured = 9.48f;
 
         float result = premiumCalculator.calculate(policy);
-        float expected = (float) 3.15;
+        float expected = 3.15f;
         assertEquals(expected, result);
     }
 }
